Added a navbar, added a form for country.
Did not get it to work with the centering of the navbar, the color only applied to the center in that case. Attempted to get the form to have center-aligned text before it's clicked and then change its alignment to left upon focus. Did not get that to work either.